% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ai_ratio}
\alias{ai_ratio}
\title{ai_ratio}
\usage{
ai_ratio(d, sr, p)
}
\arguments{
\item{d}{Standardized mean difference between subgroups}

\item{sr}{Selection ratio}

\item{p}{Proportion of the focal (minority) subgroup}
}
\value{
Adverse impact ratio
}
\description{
Convert a standardized mean subgroup difference into an adverse impact
ratio, given a selection ratio and subgroup proportion.
}
