% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbi-utils.R
\name{assert_col_vec}
\alias{assert_col_vec}
\title{assert_col_vec}
\usage{
assert_col_vec(v)
}
\arguments{
\item{v}{Intermediate variable v}
}
\value{
Refined variable v
}
\description{
Support function, refines intermediate variable for use in NBI()
}
