% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_biserial}
\alias{calc_biserial}
\title{calc_biserial}
\usage{
calc_biserial(d, p)
}
\arguments{
\item{d}{Standardized mean difference between subgroups}

\item{p}{Proportion of the focal (minority) subgroup}
}
\value{
Biserial correlation corresponding to \code{d}
}
\description{
Calculate the biserial correlation corresponding to a standardized mean
difference (d), for use in data generation with subgroup mean differences.
}
