% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_d}
\alias{calc_d}
\title{calc_d}
\usage{
calc_d(r, p)
}
\arguments{
\item{r}{Biserial correlation}

\item{p}{Proportion of the focal (minority) subgroup}
}
\value{
Standardized mean difference corresponding to \code{r}
}
\description{
Reverses \code{\link{calc_biserial}}, converting a biserial correlation
back to a standardized mean difference (d).
}
