% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbi.R
\name{combR}
\alias{combR}
\title{combR}
\usage{
combR(Rx, Ry)
}
\arguments{
\item{Rx}{Predictor inter-correlation matrix}

\item{Ry}{Predictor-criterion correlation (validity)}
}
\value{
Rxy Predictor-criterion correlation matrix
}
\description{
Support function to create predictor-criterion matrix
}
