% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbi-utils.R
\name{dimFun}
\alias{dimFun}
\title{dimFun}
\usage{
dimFun(x)
}
\arguments{
\item{x}{Input predictor weight vector}
}
\value{
x Checked and refined input predictor weight vector
}
\description{
Support function, checks input predictor weight vector x
}
