% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_weights}
\alias{get_weights}
\title{get_weights}
\usage{
get_weights(sr, props, ds, R, spac, graph = FALSE, display_solution = FALSE)
}
\arguments{
\item{sr}{Selection ratio}

\item{props}{Subgroup proportion(s)}

\item{ds}{Subgroup difference(s)}

\item{R}{Predictor/criterion correlation matrix}

\item{spac}{Number of Pareto solutions}

\item{graph}{If TRUE, plots will be generated for the Pareto-optimal curve and predictor weights}

\item{display_solution}{If TRUE, the Pareto-optimal solution will be displayed}
}
\value{
Matrix of Pareto-optimal predictor weights
}
\description{
Obtain PO predictor weights, dispatching to \code{ParetoR::ParetoR()} for
a single subgroup or \code{\link{POMA_Focal}} for multiple subgroups.
}
