% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbi-utils.R
\name{myT}
\alias{myT}
\title{myT}
\usage{
myT(x_t)
}
\arguments{
\item{x_t}{Temporary input weight vector}
}
\value{
f Temporary criterion space
}
\description{
Support function, define criterion space for intermediate step in NBI()
}
