% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbi-utils.R
\name{myTCon_eq}
\alias{myTCon_eq}
\title{myTCon_eq}
\usage{
myTCon_eq(x_t)
}
\arguments{
\item{x_t}{Temporary input weight vector}
}
\value{
ceq Temporary constraint condition
}
\description{
Support function, define constraint condition for intermediate step in NBI()
}
