% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POMPG-package.R
\docType{package}
\name{POMPG-package}
\alias{POMPG}
\alias{POMPG-package}
\title{POMPG: Pareto-Optimal Multi-Protected-Group Selection Optimization}
\description{
Extends the Pareto-optimal predictor weighting framework of the 'ParetoR' package (Normal Boundary Intersection) from a single adverse-impact protected subgroup to multiple protected subgroups simultaneously (PO-MPG). Provides POMA_Focal() for multi-group NBI optimization and supporting utilities for selection-system simulation.
}
\author{
\strong{Maintainer}: Q. Chelsea Song \email{qianqisong@gmail.com}

Authors:
\itemize{
  \item Q. Chelsea Song \email{qianqisong@gmail.com}
}

}
\keyword{internal}
