% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quiet}
\alias{quiet}
\title{quiet}
\usage{
quiet(x)
}
\arguments{
\item{x}{Expression whose printed/cat output should be suppressed}
}
\value{
The (invisible) result of evaluating \code{x}
}
\description{
Suppress function outputs written via \code{cat}/\code{print}. Acquired
from: http://r.789695.n4.nabble.com/Suppressing-output-e-g-from-cat-td859876.html
}
