% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbi-utils.R
\name{Weight_Generate}
\alias{Weight_Generate}
\title{Weight_Generate}
\usage{
Weight_Generate(n, k)
}
\arguments{
\item{n}{Number of objects (i.e., number of predictor and criterion)}

\item{k}{Number of Pareto points}
}
\value{
Weight_Generate
}
\description{
Function intended to test the weight generation scheme for NBI for > 2 objectives
}
